aerosync-ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
openai>=1.30.0
pyyaml>=6.0.1
httpx>=0.27.0
